📝 Refactor : Edit package path & Remove duplicate code
Refactor : Edit package path & Remove duplicate code
