+++ src/app/visitor/VisitorManager.java
@@ -8,14 +8,14 @@ import java.util.Map;
 import java.util.Scanner;
 
 import app.animal.AnimalEnum.Species;
-import app.common.id.IdGeneratorUtil;
+import app.common.IdGeneratorUtil;
 
 public class VisitorManager {
 
 	Scanner in = new Scanner(System.in);
 	Map<String, Reservation> reservations = new HashMap<>();
 
-	String id; // 예약 번호 
+	String id; // 예약 번호
 	String name;
 	String phone;
 	String date;
@@ -58,7 +58,7 @@ public class VisitorManager {
 
 	public void zooInformation() {
 
-		// 동물원 정보 
+		// 동물원 정보
 		System.out.println("이름 : Manazoo 동물원");
 		System.out.println("주소 : 서울시 종각");
 		System.out.println("문의전화 : 02)123-1234 \n");
@@ -71,7 +71,7 @@ public class VisitorManager {
 		System.out.println(); // 줄 바꾸기
 		System.out.println(); // 줄 바꾸기
 
-		// 가격 정보 
+		// 가격 정보
 		System.out.printf("대인 티켓 : %d원 \n", ADULT_PRICE);
 		System.out.printf("소인 티켓 : %d원 \n", CHILD_PRICE);
 		System.out.println();
@@ -79,18 +79,18 @@ public class VisitorManager {
 
 	public void reservation() {
 		while (true) {
-			// 예약번호 생성 
+			// 예약번호 생성
 			id = IdGeneratorUtil.generateId();
 
-			// 예약 정보 입력 받기 
+			// 예약 정보 입력 받기
 			inputInformation();
 
 			// 예약 정보 확인 받고, 결제 여부 결정
 			if (confirmInformation()) {
-				break; // while문 깨고 payment() 진행 
+				break; // while문 깨고 payment() 진행
 			}
 		}
-		// 결제 진행 
+		// 결제 진행
 		payment();
 	}
 
@@ -98,14 +98,12 @@ public class VisitorManager {
 		System.out.println("예약 정보를 입력해 주세요.");
 
 		// 방문날짜 입력 받기
-		DateTimeFormatter formatter = DateTimeFormatter
-				.ofPattern("yyyy-MM-dd");
+		DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 		while (true) {
 			System.out.println("방문 일시 (YYYY-MM-DD) : ");
 			String inDate = in.nextLine();
 			try {
-				LocalDate localDate = LocalDate.parse(inDate,
-						formatter);
+				LocalDate localDate = LocalDate.parse(inDate, formatter);
 				date = localDate.format(formatter);
 				break; // 성공적으로 파싱되면 루프 종료
 			} catch (DateTimeParseException e) {
@@ -163,19 +161,18 @@ public class VisitorManager {
 		System.out.println("방문 일시 : " + date);
 		System.out.println("대인 인원수 : " + adultCount);
 		System.out.println("소인 인원수 : " + childCount);
-		totalPrice = adultCount * ADULT_PRICE
-				+ childCount * CHILD_PRICE;
+		totalPrice = adultCount * ADULT_PRICE + childCount * CHILD_PRICE;
 		System.out.printf("금액 : %d 원 \n", totalPrice);
 		while (true) {
 
-			// 예약 정보 확정 받고, 결제 여부 결정 
+			// 예약 정보 확정 받고, 결제 여부 결정
 			System.out.println("1.다시 입력 2.결제");
 			String answer = in.nextLine();
 			if (answer.equals("1")) {
 				// < 다시입력 > (내부 while문 깨고, 외부 while로 반복)
 				break;
 			} else if (answer.equals("2")) {
-				return true; // 결제로 이동 
+				return true; // 결제로 이동
 			} else {
 				System.out.println("잘못된 선택입니다.");
 			}
@@ -191,7 +188,7 @@ public class VisitorManager {
 			System.out.println("이름: ");
 			name = in.nextLine();
 
-			// 전화번호 : 지정된 형식으로 받기 
+			// 전화번호 : 지정된 형식으로 받기
 			String phonePattern = "^\\d{3}-\\d{4}-\\d{4}$";
 			while (true) {
 				System.out.println("폰 번호 : ");
@@ -204,20 +201,18 @@ public class VisitorManager {
 				}
 			}
 
-			// 결제 정보 확인 받고, 결제 진행 여부 결정 
+			// 결제 정보 확인 받고, 결제 진행 여부 결정
 			System.out.printf("결제 금액 : %d 원 \n", totalPrice);
 			System.out.println("결제하시겠습니까?");
 			System.out.println("1.예 2.아니오(나가기)");
 			String answer = in.nextLine();
 			if (answer.equals("1")) {
 
-				// 예약(Reservation) 객체 생성하여 Map에 저장 
-				Reservation reservation = new Reservation(id, name,
-						phone, date, adultCount, childCount,
-						totalPrice);
+				// 예약(Reservation) 객체 생성하여 Map에 저장
+				Reservation reservation = new Reservation(id, name, phone, date, adultCount, childCount, totalPrice);
 				reservations.put(id, reservation);
 
-				// 예약 및 결제 성공 메시지 
+				// 예약 및 결제 성공 메시지
 				System.out.println("예약 및 결제 성공!");
 				System.out.println(reservation);
 				System.out.println();
@@ -235,20 +230,19 @@ public class VisitorManager {
 	public void viewReservation() {
 		while (true) {
 
-			// 예약 목록이 없을 경우 
+			// 예약 목록이 없을 경우
 			if (reservations.isEmpty()) {
 				System.out.println("(예약 목록 없음)");
 				return;
 			} else {
 
-				// 예약번호로 검색 
+				// 예약번호로 검색
 				System.out.println("예약번호 : ");
 				String findId = in.nextLine();
 				if (reservations.containsKey(findId)) {
 
-					// 검색된 예약 내용 보여주기 
-					Reservation reservation = reservations
-							.get(findId);
+					// 검색된 예약 내용 보여주기
+					Reservation reservation = reservations.get(findId);
 					System.out.println(reservation);
 					return;
 				} else {
@@ -261,32 +255,29 @@ public class VisitorManager {
 	public void changeReservationDate() {
 		while (true) {
 
-			// 예약 목록이 없을 경우 
+			// 예약 목록이 없을 경우
 			if (reservations.isEmpty()) {
 				System.out.println("(예약 목록 없음)");
 				return;
 			} else {
 
-				// 예약번호로 검색 
+				// 예약번호로 검색
 				System.out.println("예약번호 : ");
 				String findId = in.nextLine();
 				if (reservations.containsKey(findId)) {
 
-					// 검색된 예약의 방문일정 보여주기 
-					Reservation reservation = reservations
-							.get(findId);
-					System.out.println(
-							"현재 방문일정 : " + reservation.getDate());
+					// 검색된 예약의 방문일정 보여주기
+					Reservation reservation = reservations.get(findId);
+					System.out.println("현재 방문일정 : " + reservation.getDate());
 
 					// 변경할 일정 입력 받기
 					System.out.println("변경 일정 (YYYY-MM-DD) : ");
 					String inDate = in.nextLine();
 
-					// 일정 변경하기 
+					// 일정 변경하기
 					reservation.setDate(inDate);
 					System.out.println("방문일정 변경 성공!");
-					System.out.println(
-							"방문일정 : " + reservation.getDate());
+					System.out.println("방문일정 : " + reservation.getDate());
 					return;
 				} else {
 					System.out.println("예약번호를 다시 입력해 주세요. ");
@@ -298,20 +289,19 @@ public class VisitorManager {
 	public void cancelReservation() {
 		while (true) {
 
-			// 예약 목록이 없을 경우 
+			// 예약 목록이 없을 경우
 			if (reservations.isEmpty()) {
 				System.out.println("(예약 목록 없음)");
 				return;
 			} else {
 
-				// 예약번호로 검색 
+				// 예약번호로 검색
 				System.out.println("예약번호 : ");
 				String findId = in.nextLine();
 				if (reservations.containsKey(findId)) {
 
-					// 검색된 예약 내용 보여주기 
-					Reservation reservation = reservations
-							.get(findId);
+					// 검색된 예약 내용 보여주기
+					Reservation reservation = reservations.get(findId);
 					System.out.println(reservation);
 
 					System.out.println("예약을 취소하시겠습니까?");
@@ -319,13 +309,13 @@ public class VisitorManager {
 					String answer = in.nextLine();
 					if (answer.equals("1")) {
 
-						// 예약 취소, Map에서 데이터 삭제 
+						// 예약 취소, Map에서 데이터 삭제
 						reservations.remove(findId);
 						System.out.println("예약 취소 완료");
 						return;
 					} else if (answer.equals("2")) {
 
-						// 2.아니오(나가기) 선택 
+						// 2.아니오(나가기) 선택
 						return;
 					} else {
 						System.out.println("잘못된 선택입니다.");
